Work in progress on Minecraft 1.18 platform support Work in progress on supporting negative heights

+++ WorldPainter/WPGUI/src/main/java/org/pepsoft/worldpainter/painting/DimensionPainter.java
@@ -253,7 +253,7 @@ public final class DimensionPainter {
 
                                 @Override
                                 boolean isFilled(int x, int y) {
-                                    return dimension.getLayerValueAt(layer, x, y) == ((DiscreteLayerPaint) paint).getValue();
+                                    return dimension.getLayerValueAt(layer, x, y) == (int) ((DiscreteLayerPaint) paint).getValue();
                                 }
                             };
                         }
